// src/shared/enums/error-code.enum.ts
export enum ErrorCodeEnum {
  AUTH_INVALID_CREDENTIALS = 'AUTH_INVALID_CREDENTIALS',
  AUTH_TOKEN_EXPIRED = 'AUTH_TOKEN_EXPIRED',
  AUTH_UNAUTHORIZED = 'AUTH_UNAUTHORIZED',

  USER_NOT_FOUND = 'USER_NOT_FOUND',
  USER_ALREADY_EXISTS = 'USER_ALREADY_EXISTS',
  USER_EMAIL_ALREADY_USED = 'USER_EMAIL_ALREADY_USED',

  BRAND_NOT_FOUND = 'BRAND_NOT_FOUND',
  BRAND_ALREADY_EXIST = 'BRAND_ALREADY_EXIST',

  CATEGORY_NOT_FOUND = 'BRAND_NOT_FOUND',
  CATEGORY_ALREADY_EXIST = 'BRAND_ALREADY_EXIST',

  PRODUCT_NOT_FOUND = 'PRODUCT_NOT_FOUND',
  PRODUCT_NOT_AVAILABLE = 'PRODUCT_NOT_AVAILABLE',

  PRODUCT_VARIANT_OUT_OF_STOCK = 'PRODUCT_OUT_OF_STOCK',
  PRODUCT_VARIANT_NOT_FOUND = 'PRODUCT_VARIANT_NOT_FOUND',

  ORDER_NOT_FOUND = 'ORDER_NOT_FOUND',
  ORDER_PENDING_NOT_FOUND = 'ORDER_PENDING_NOT_FOUND',
  ORDER_EMPTY = 'ORDER_EMPTY',
  ORDER_ACCESS_DENIED = 'ORDER_ACCESS_DENIED',
  ORDER_ALREADY_PAID = 'ORDER_ALREADY_PAID',
  ORDER_INVALID_STATUS = 'ORDER_INVALID_STATUS',

  PAYMENT_NOT_FOUND = 'PAYMENT_NOT_FOUND',
  PAYMENT_FAILED = 'PAYMENT_FAILED',
  PAYMENT_ALREADY_PROCESSED = 'PAYMENT_ALREADY_PROCESSED',
  PAYMENT_INVALID_STATUS = 'PAYMENT_INVALID_STATUS',
}